PC(backend - reconhecimento de pessoas )/src/requirements.txt
ultralytics
mediapipe
opencv-python
flask
paho-mqtt
python-dotenv
supabase
requests
